Feat 🎉: Remodel the decks listing page.
We added the feature to add a deck inside another deck recursively.
Changed the decks listing page to a column list of decks.
Added a drag and drop feature, to easily insert a deck inside another deck or re-arange decks.

Missing: Add the functionality to actually add a deck inside another deck.